chore: split script into modules
