+++ js/common.js
@@ -3,7 +3,7 @@
 // -------------------------
 // Caricamento dati e popolamento dropdown
 // -------------------------
-function loadDropdownData(jsonPath, selectId, key) {
+export function loadDropdownData(jsonPath, selectId, key) {
   fetch(jsonPath)
     .then(response => response.json())
     .then(data => {
@@ -21,7 +21,7 @@ function loadDropdownData(jsonPath, selectId, key) {
     .catch(error => handleError(`Errore caricando ${jsonPath}: ${error}`));
 }
 
-function populateDropdown(selectId, options) {
+export function populateDropdown(selectId, options) {
   const select = document.getElementById(selectId);
   if (!select) {
     handleError(`Elemento #${selectId} non trovato!`);
@@ -37,128 +37,9 @@ function populateDropdown(selectId, options) {
 }
 
 // -------------------------
-// Conversione dei dati della razza (raw JSON → formato standard)
-// -------------------------
-function convertRaceData(rawData) {
-  // Size
-  let size = "Unknown";
-  if (Array.isArray(rawData.size)) {
-    size = (rawData.size[0] === "M") ? "Medium" : (rawData.size[0] === "S") ? "Small" : rawData.size[0];
-  } else {
-    size = rawData.size || "Unknown";
-  }
-
-  // Speed
-  let speed = {};
-  if (rawData.speed) {
-    if (typeof rawData.speed === "object") {
-      for (let key in rawData.speed) {
-        speed[key] = (typeof rawData.speed[key] === "boolean")
-          ? (key === "fly" ? "equal to your walking speed" : "unknown")
-          : rawData.speed[key];
-      }
-    } else {
-      speed = rawData.speed;
-    }
-  }
-
-  // Senses
-  let senses = {};
-  if (rawData.senses && typeof rawData.senses === "object") {
-    senses = rawData.senses;
-  } else if (rawData.darkvision) {
-    senses.darkvision = rawData.darkvision;
-  }
-
-  // Ability Bonus (minimale gestione)
-  let ability_bonus = { options: [] };
-  if (rawData.ability && Array.isArray(rawData.ability)) {
-    rawData.ability.forEach(ability => {
-      if (ability.choose && ability.choose.weighted) {
-        const weights = ability.choose.weighted.weights;
-        if (weights.length === 2 && weights.includes(2)) {
-          ability_bonus.options.push({ type: "fixed", values: { any: 2, any_other: 1 } });
-        } else if (weights.length === 3) {
-          ability_bonus.options.push({ type: "three", values: { any: 1, any_other: 1, any_other_2: 1 } });
-        }
-      }
-    });
-  }
-
-  // Tratti
-  let traits = [];
-  const rawEntries = rawData.entries || [];
-  rawEntries.forEach(entry => {
-    if (entry.name && entry.entries) {
-      // Un semplice concatenamento dei testi delle entry
-      const description = Array.isArray(entry.entries) ? entry.entries.join(" ") : entry.entries;
-      traits.push({
-        name: entry.name,
-        description: description,
-        level_requirement: 1
-      });
-    }
-  });
-
-  // Lingue
-  let languages = { fixed: [], choice: 0, options: [] };
-  if (rawData.languageProficiencies && rawData.languageProficiencies.length > 0) {
-    const lp = rawData.languageProficiencies[0];
-    for (let lang in lp) {
-      if (lp[lang] === true) {
-        languages.fixed.push(lang.charAt(0).toUpperCase() + lang.slice(1));
-      } else if (typeof lp[lang] === "number") {
-        languages.choice = lp[lang];
-      }
-    }
-    if (languages.choice > 0 && languages.options.length === 0) {
-      languages.options.push("Any other language you and your DM agree is appropriate");
-    }
-  }
-
-  // Skill Choices
-  let skill_choices = null;
-  if (rawData.skillProficiencies && rawData.skillProficiencies.length > 0) {
-    const sp = rawData.skillProficiencies[0].choose;
-    if (sp && sp.from) {
-      const count = sp.count ? sp.count : 1;
-      skill_choices = { number: count, options: sp.from };
-    }
-  }
-
-  // Tool Choices
-  let tool_choices = null;
-  if (rawData.toolProficiencies && Array.isArray(rawData.toolProficiencies)) {
-    rawData.toolProficiencies.forEach(tp => {
-      if (tp.choose && tp.choose.from) {
-        tool_choices = { number: 1, options: tp.choose.from };
-      }
-    });
-  }
-
-  // Spellcasting (qui manteniamo la struttura originale, che potrà essere ulteriormente elaborata)
-  let spellcasting = rawData.additionalSpells || null;
-
-  return {
-    name: rawData.name,
-    source: rawData.source + (rawData.page ? `, page ${rawData.page}` : ""),
-    size: size,
-    speed: speed,
-    senses: senses,
-    ability_bonus: ability_bonus,
-    traits: traits,
-    rawEntries: rawEntries,
-    spellcasting: spellcasting,
-    languages: languages,
-    skill_choices: skill_choices,
-    tool_choices: tool_choices
-  };
-}
-
-// -------------------------
 // Funzione per caricare le lingue da un file JSON
 // -------------------------
-function loadLanguages(callback) {
+export function loadLanguages(callback) {
   fetch("data/languages.json")
     .then(response => response.json())
     .then(data => {
@@ -172,49 +53,9 @@ function loadLanguages(callback) {
 }
 
 // -------------------------
-// Helper per estrarre il nome di uno spell (ricorsivamente)
-// -------------------------
-function extractSpellName(data) {
-  if (Array.isArray(data)) {
-    if (typeof data[0] === "string") {
-      return data[0].split("#")[0];
-    }
-  } else if (typeof data === "object") {
-    for (let key in data) {
-      const result = extractSpellName(data[key]);
-      if (result) return result;
-    }
-  }
-  return null;
-}
-
-// -------------------------
-// Filtro incantesimi (es. "level=0|class=Wizard")
-// -------------------------
-function filterSpells(spells, filterString) {
-  const conditions = filterString.split("|");
-  return spells.filter(spell => {
-    let valid = true;
-    conditions.forEach(cond => {
-      const parts = cond.split("=");
-      if (parts.length === 2) {
-        const key = parts[0].trim().toLowerCase();
-        const value = parts[1].trim().toLowerCase();
-        if (key === "level") {
-          if (parseInt(spell.level) !== parseInt(value)) valid = false;
-        } else if (key === "class") {
-          if (!spell.spell_list.map(x => x.toLowerCase()).includes(value)) valid = false;
-        }
-      }
-    });
-    return valid;
-  });
-}
-
-// -------------------------
 // Funzione per mostrare un messaggio d'errore
 // -------------------------
-function handleError(message) {
+export function handleError(message) {
   console.error("❌ " + message);
   alert("⚠️ " + message);
 }
@@ -222,7 +63,7 @@ function handleError(message) {
 // -------------------------
 // Funzione per il rendering di tabelle (es. per alcune entry)
 // -------------------------
-function renderTables(entries) {
+export function renderTables(entries) {
   let html = "";
   if (!entries || !Array.isArray(entries)) return html;
   entries.forEach(entry => {
