+++ js/raceData.js
@@ -1,0 +1,122 @@
+export function convertRaceData(rawData) {
+  // Size
+  let size = "Unknown";
+  if (Array.isArray(rawData.size)) {
+    size = (rawData.size[0] === "M") ? "Medium" : (rawData.size[0] === "S") ? "Small" : rawData.size[0];
+  } else {
+    size = rawData.size || "Unknown";
+  }
+
+  // Speed
+  let speed = {};
+  if (rawData.speed) {
+    if (typeof rawData.speed === 'object') {
+      for (let key in rawData.speed) {
+        speed[key] = (typeof rawData.speed[key] === 'boolean')
+          ? (key === 'fly' ? 'equal to your walking speed' : 'unknown')
+          : rawData.speed[key];
+      }
+    } else {
+      speed = rawData.speed;
+    }
+  }
+
+  // Senses
+  let senses = {};
+  if (rawData.senses && typeof rawData.senses === 'object') {
+    senses = rawData.senses;
+  } else if (rawData.darkvision) {
+    senses.darkvision = rawData.darkvision;
+  }
+
+  // Ability Bonus
+  let ability_bonus = { options: [] };
+  if (rawData.ability && Array.isArray(rawData.ability)) {
+    rawData.ability.forEach(ability => {
+      if (ability.choose && ability.choose.weighted) {
+        const weights = ability.choose.weighted.weights;
+        if (weights.length === 2 && weights.includes(2)) {
+          ability_bonus.options.push({ type: 'fixed', values: { any: 2, any_other: 1 } });
+        } else if (weights.length === 3) {
+          ability_bonus.options.push({ type: 'three', values: { any: 1, any_other: 1, any_other_2: 1 } });
+        }
+      }
+    });
+  }
+
+  // Variant Feature and Traits
+  let variant_feature_choices = null;
+  let traits = [];
+  const rawEntries = rawData.entries || [];
+  rawEntries.forEach(entry => {
+    if (entry.type === 'inset' && entry.name && entry.name.toLowerCase().includes('variant feature')) {
+      let variantText = '';
+      if (Array.isArray(entry.entries)) {
+        variantText = entry.entries.map(opt => {
+          let optDesc = Array.isArray(opt.entries) ? opt.entries.join(' ') : opt.entries;
+          return `${opt.name}: ${optDesc}`;
+        }).join(' | ');
+      }
+      traits.push({
+        name: entry.name,
+        description: variantText,
+        level_requirement: 1
+      });
+    }
+  });
+
+  // Languages
+  let languages = { fixed: [], choice: 0, options: [] };
+  if (rawData.languageProficiencies && rawData.languageProficiencies.length > 0) {
+    const lp = rawData.languageProficiencies[0];
+    for (let lang in lp) {
+      if (lp[lang] === true) {
+        languages.fixed.push(lang.charAt(0).toUpperCase() + lang.slice(1));
+      } else if (typeof lp[lang] === 'number') {
+        languages.choice = lp[lang];
+      }
+    }
+    if (languages.choice > 0 && languages.options.length === 0) {
+      languages.options.push('Any other language you and your DM agree is appropriate');
+    }
+  }
+
+  // Skill Choices
+  let skill_choices = null;
+  if (rawData.skillProficiencies && rawData.skillProficiencies.length > 0) {
+    const sp = rawData.skillProficiencies[0].choose;
+    if (sp && sp.from) {
+      const count = sp.count ? sp.count : 1;
+      skill_choices = { number: count, options: sp.from };
+    }
+  }
+
+  // Tool Choices
+  let tool_choices = null;
+  if (rawData.toolProficiencies && Array.isArray(rawData.toolProficiencies)) {
+    rawData.toolProficiencies.forEach(tp => {
+      if (tp.choose && tp.choose.from) {
+        tool_choices = { number: 1, options: tp.choose.from };
+      }
+    });
+  }
+
+  // Spellcasting (if present)
+  let spellcasting = rawData.additionalSpells || null;
+
+  return {
+    name: rawData.name,
+    source: rawData.source + (rawData.page ? `, page ${rawData.page}` : ''),
+    size: size,
+    speed: speed,
+    senses: senses,
+    ability_bonus: ability_bonus,
+    traits: traits,
+    rawEntries: rawEntries,
+    spellcasting: spellcasting,
+    languages: languages,
+    skill_choices: skill_choices,
+    tool_choices: tool_choices,
+    variant_feature_choices
+  };
+}
